==> notebooks/01-Data_Cleaning.ipynb <==
# Major League Baseball Home Run Exit Velocities

One of the most recent debates in Major League Baseball is focused on why more home runs were hit in the 2017 season than any other season in the leagues history. There were 6,105 home runs hit in the 2017 season, which is more home runs than in the peak of the steroid era and everyone wants to know why. One specific area of interest for investigation is related to the most important item to the game, the baseball. During the 2017 season there were numerous complaints from major league pitchers that the ball felt different and the result was a record breaking year for home runs.

Thanks to Baseball Savant and their StatCast technology we can begin to understand what is happening by analyzing the data tracked on every pitch that was hit for a home run. After a high level review of the data and some outside research I found a specific stat known as the Home Run Exit Velocity which is defined by Baseball Savant as the speed at which the baseball leaves the bat after being hit. Home Run Exit Velocity will be the target variable for my investigation into what is causing an increase in home runs.

Using machine learning techniques I will build a production level regression model to draw a conclusion on what features involved in a home run are most influential to the home run exit velocity. The data that will be used to train this model will come from three different aspects:
1. StatCast pitch data for all home run hits
2. Data for a sample of baseballs used in the 2015, 2016 and 2017 seasons
3. Baseball player personal stats (height, weight, age)

import pandas as pd
import numpy as np
from scipy import stats 

from bs4 import BeautifulSoup

import requests
import json
import time

import warnings
warnings.filterwarnings('ignore')

## Data Cleaning

### Import the Data

The following notebook will load in two datasets for review and cleaning:
1. Each home run hit during the 2015, 2016 and 2017 MLB seasons
2. Stats from a sample of the actual baseballs taken from the 2015, 2016 and 2017 MLB seasons
3. I will also perform a webscrape of the website Baseball Savant for each players personal physical stats that hit a home run in the 2015, 2016 and 2017 seasons.
    - Height
    - Weight
    - Age
    
URL : https://baseballsavant.mlb.com/

#### Home Runs

To start I will review and clean the largest of my three datasets which is the statistical data on each home run hit during the 2015, 2016 and 2017 seasons.

hr_df = pd.read_csv('../data/home_runs_15_16_17.csv')
baseballs_df = pd.read_excel('../data/baseballs.xlsx')

hr_df.columns = [x.lower().replace(' ', '_') for x in hr_df.columns]

hr_df.head()

hr_df.info()

#### Remove unnecessary features (noise)

After a review of all the available features within this dataset the following will be dropped:

**NOTE:** An in depth review of each of the columns can be found at basball savant: https://baseballsavant.mlb.com/

- Dropped because of lack of data:
    - `spin_dir`
    - `spin_rate_deprecated`
    - `break_angle_deprecated`
    - `break_length_deprecated` 
    - `hit_location`
    - `tfs_deprecated`
    - `tfs_zulu_deprecated`
    - `iso_value`
    - `launch_speed_angle`: the values of 1-6 are unbalanced and I cannot find the exact meanings. I already have seperate features for both launch angle and launch speed.
- Dropped because of no relation to the actual home run stats (launch speed, launch angle, etc.)
    - `on_3b`
    - `on_2b`
    - `on_1b`
    - `umpire`
    - `events`
    - `description`
    - `des`
    - `game_type`
    - `stand`
    - `type`
    - `balls`
    - `strikes`
    - `outs_when_up`
    - `inning`
    - `inning_topbot`
    - `fielder_2`
    - `sv_id`
    - `pitcher.1`
    - `fielder_2.1`
    - `fielder_3`
    - `fielder_4`
    - `fielder_5`
    - `fielder_6`
    - `fielder_7` 
    - `fielder_8 `
    - `fielder_9`
    - `estimated_ba_using_speedangle`
    - `estimated_woba_using_speedangle`
    - `woba_value`
    - `woba_denom` 
    - `babip_value`
    - `at_bat_number`
    - `pitch_number `
    - `home_score` 
    - `away_score` 
    - `bat_score`
    - `fld_score`
    - `post_away_score`
    - `post_home_score`
    - `post_bat_score`
    - `post_fld_score`
    - `if_fielding_alignment `
    - `of_fielding_alignment`
    - `game_pk`
    - `pitch_name` (duplicate to pitch type)
    - `home_team`
    - `away_team`
    - `hc_x : coordinate of where the ball is fielded`
    - `hc_y : coordinate of where the ball is fielded`

col = [ 
    'umpire',
    'spin_dir',
    'spin_rate_deprecated',
    'break_angle_deprecated',
    'break_length_deprecated',
    'events',
    'description', 
    'des',
    'game_type', 
    'stand',
    'type', 
    'hit_location', 
    'balls', 
    'strikes', 
    'on_3b', 
    'on_2b', 
    'on_1b', 
    'outs_when_up',
    'inning',
    'inning_topbot',
    'tfs_deprecated', 
    'tfs_zulu_deprecated', 
    'fielder_2', 
    'sv_id', 
    'pitcher.1', 
    'fielder_2.1', 
    'fielder_3', 
    'fielder_4', 
    'fielder_5', 
    'fielder_6', 
    'fielder_7', 
    'fielder_8', 
    'fielder_9', 
    'estimated_ba_using_speedangle', 
    'estimated_woba_using_speedangle', 
    'woba_value', 
    'woba_denom', 
    'babip_value', 
    'at_bat_number', 
    'pitch_number', 
    'home_score', 
    'away_score', 
    'bat_score', 
    'fld_score', 
    'post_away_score', 
    'post_home_score', 
    'post_bat_score', 
    'post_fld_score', 
    'if_fielding_alignment', 
    'of_fielding_alignment',
    'iso_value',
    'game_pk',
    'pitch_name',
    'away_team',
    'home_team',
    'hc_x',
    'hc_y',
    'launch_speed_angle'
]

hr_df.drop(col, axis=1, inplace=True)

#### Removing nulls

One of the features that is a very important aspect to a specific players home run stats is the pitch in which they hit a home run. So, I will be dropping the rows in which have nulls documented within the pitch type that was thrown.

col=['pitch_type']
hr_df.dropna(subset=col, inplace=True)

#### Updating datatypes

Ensuring specific data types are captured in the correct format

hr_df['game_date'] = pd.to_datetime(hr_df['game_date'])

hr_df.info()

hr_df['pitch_type'].isnull().sum()

#### Filling in Nulls (Pitch Stats)

A lot of the pitching data related to the actual pitch that was thrown is missing. Using a groupby and a function I can fill these values in for the specific pitcher that threw the pitch. 

Steps to fill in these values:
1. Create a dataframe grouped by pitcher and pitch type to get an average for each pitch statistic
2. Create a function that will look up this picher by their ID and then look up their average stat from the newly created dataframe to fill in whatever missing values are missing for the specific pitch record
3. Apply the newly created function to the pitching features where nulls were identified

hr_df.isnull().sum()[['release_speed', 
                      'release_pos_x', 
                      'release_pos_y',
                      'release_pos_z',
                      'release_spin_rate',
                      'release_extension',
                      'pfx_x',
                      'pfx_z',
                      'plate_x',
                      'vx0',
                      'vy0',
                      'vz0',
                      'ax',
                      'ay',
                      'az',
                      'effective_speed']]

pitch_avg = hr_df.groupby(['pitcher', 'pitch_type'])[['release_speed', 
                                                      'release_pos_x', 
                                                      'release_pos_y',
                                                      'release_pos_z',
                                                      'release_spin_rate',
                                                      'release_extension',
                                                      'pfx_x',
                                                      'pfx_z',
                                                      'plate_x',
                                                      'vx0',
                                                      'vy0',
                                                      'vz0',
                                                      'ax',
                                                      'ay',
                                                      'az',
                                                      'effective_speed']].mean()

pitch_avg.reset_index(inplace=True)

pitch_avg.head()

def pitcher_avg(pitch, col):
    return float(pitch_avg[(pitch_avg['pitcher']==pitch.pitcher) & 
                           (pitch_avg['pitch_type']==pitch.pitch_type)][col])

def pitch_stats_apply(columns):
    for col in columns:
        hr_df[col] = hr_df.apply(lambda x: pitcher_avg(x, col) if pd.isnull(x[col]) else x[col],1)
    return hr_df

columns = pitch_avg
hr_df = pitch_stats_apply(columns)

hr_df['zone'].iloc[9248] = 4

hr_df['zone'].iloc[9338] = 4

hr_df.info()

#### Filling in Nulls (Home Run Hit Stats)

With as many nulls as possible now filled in for the missing pitch stats I will now perform the same steps for the missing home run hit stats. 

Steps to fill in these values:
1. Create a dataframe grouped by batter, the pitch type and the batted ball type to get an average for each for the batting statistics that contain nulls
2. Create a function that will look up the batter by their ID and then look up their average stat from the newly created dataframe to fill in missing values
3. Apply the newly created function to the batter features where nulls were identified

batter_stats = hr_df.groupby(['pitch_type', 'batter', 'bb_type']).mean()[['hit_distance_sc']]

batter_stats.reset_index(inplace=True)

batter_stats.head()

def batter_avg(pitch, col):
    return float(batter_stats[(batter_stats['batter']==pitch.batter) &
                              (batter_stats['pitch_type']==pitch.pitch_type) &
                              (batter_stats['bb_type']==pitch.bb_type)][col])

def batter_stats_apply(columns):
    for col in columns:
        hr_df[col] = hr_df.apply(lambda x: batter_avg(x, col) 
                                               if pd.isnull(x[col])
                                               else x[col],1)

columns = batter_stats
batter_stats_apply(columns)

hr_df.info()

#### Final missing value fill

I have now filled in as many nulls as possible within the dataset and will now drop all remaining null values

hr_df.dropna(inplace=True)

len(hr_df)

### Save the Data

hr_df.to_csv('../data/final_clean_hr.csv')

### Baseball Data

Next, I will review and clean the sampled baseballs from the 2015, 2016 and 2017 MLB seasons

baseballs_df.columns = [i.lower().replace(' ', '_') for i in baseballs_df.columns]

baseballs_df.head()

#### Dropping unnecessary features (noise)
- Dropping because of their irrelevance to home run statistics
    - ball code
    - sn
- Dropping because of their lack of data
    - #_of_good_shots_before_damage

col = [
    'ball_code',
    '#_of_good_shots_before_damage',
    'sn'
]

baseballs_df.drop(col, axis=1, inplace=True)

baseballs_df.info()

#### Updating datatypes 

Updating fetures to their correct data types and updating the index for this dataframe to datetime for efficiency

baseballs_df['year'] = pd.to_datetime(baseballs_df['year'])

baseballs_df.set_index('year', inplace=True)

baseballs_df.head()

#### Creating a final average of the baseballs to use for each MLB season

- 2015
    - Get the average of each statistic from the 2015 season and create as a dataframe
- 2016
    - Get the average of each statistic from the 2016 season and create as a dataframe
- 2017
    - It was extremely difficult to find the data for the baseballs from the 2017 season. Using a couple of articles online I was able to construct some averages for 2017
        - Sources: 
            - https://www.theringer.com/2017/6/14/16044264/2017-mlb-home-run-spike-juiced-ball-testing-reveal-155cd21108bc 
            - https://fivethirtyeight.com/features/juiced-baseballs/

bb_15 = baseballs_df['2015'][['weight_(oz)',
                      'circumference_(in)',
                      'avg_seam_height',
                      'std_of_seam_height',
                      'avg_ccor',
                      'avg_ds']].mean().to_frame().T
bb_15['year'] = 2015
bb_15

bb_16 = baseballs_df['2016'][['weight_(oz)',
                      'circumference_(in)',
                      'avg_seam_height',
                      'std_of_seam_height',
                      'avg_ccor',
                      'avg_ds']].mean().to_frame().T
bb_16['year'] = 2016
bb_16

mlb_baseballs = bb_15.append(bb_16).reset_index()
mlb_baseballs.drop('index', axis=1, inplace=True)
mlb_baseballs.drop('std_of_seam_height', axis=1, inplace=True)
mlb_baseballs

#### The 2017 baseballs

The largest point of emphasis that was examined within the baseballs was the weight and ccor (cylindrical coefficient of restitution). Because these are the main data points I will focus on updating these values for the 2017 baseballs.

"The newer cores weigh about a half a gram less than the older ones, which might be enough to cause baseballs hit on a typical home run trajectory to fly about 6 inches farther." -FiveThirtyEight

- Weight
        - 5.12250 - 0.017637 (.5 grams converted to ouces) = 5.104863
- Avg ccor (Ccylindrical coefficient of restitution) "The bounciness of the baseball aka "The Pill""
        - 0.494100 - 0.017637 (.5 grams converted to ouces) = .476463

bb_17 = pd.DataFrame(
    [[.476463,.038,9.08,5.104863,12926,2017]], columns=['avg_ccor', 
                                                   'avg_seam_height',
                                                   'circumference_(in)',
                                                   'weight_(oz)',
                                                   'avg_ds',
                                                   'year']
)
bb_17

mlb_baseballs = mlb_baseballs.append(bb_17, sort=False).reset_index().drop('index', axis=1)

mlb_baseballs

### Save the Data

mlb_baseballs.to_csv('../data/final_mlb_baseballs.csv')

### Player Personal Stats

Finally, I want to include each batters personal stats to draw a conclusion about whether or not a players size and weight are contributing to their home runs exit velocities.

#### Webscapre Baseball Savant

In order to scrape a website an appropriate status code must be received. Using a player at random I will send a request to the website to ensure a successful connection.
- I want to receive a status code of 200

url = "https://baseballsavant.mlb.com/savant-player/j-d-martinez-502110?stats=career-r-hitting-mlb"
res =  requests.get(url)

if res.status_code == 200:
    print(f'Successfull status code received : {res.status_code}')

#### The webscrape

Lets ensure we know how many unique players we'll need to scrape. 

There are 751 players that have hit homes runs since 2015. I will scrape each of these players personal stats:
- Height
- Weight
- Age

batter_id = hr_df['batter'].unique()

len(batter_id)

#### Running the scrape

Using Pythons Beautiful Soup Library I will run a loop through each players personal page on the Baseball Savant site and obtain their speicific personal stats.

def get_batters_info(batter_id):
    attrs = {
        'batter_id': [str(x) for x in batter_id],
        'height':[],
        'weight':[],
        'age':[],
        'position':[],
        'bats':[],
        'throws':[]
    }

    attr_names = [
        'height',
        'weight',
        'age',
        'position',
        'bats',
        'throws'
    ]

    for idx, b in enumerate(batter_id):
        url = "https://baseballsavant.mlb.com/savant-player/{}?stats=career-r-hitting-mlb".format(b)
        res =  requests.get(url)
        soup = BeautifulSoup(res.content, 'lxml')
        for col_name, attribute in zip(attr_names, soup.find_all('div', {'class':'box-text'})):
            attrs[col_name].append(str(attribute.text))
        if idx % 50 == 0:
            with open('../data/baseball_player7.json', 'w') as f:
                json.dump(attrs, f)
        time.sleep(2)
    
    with open('../data/baseball_player7.json', 'w') as f:
        json.dump(attrs, f)
    
    return attrs

#### Webscrape results

The below line of code has been commented out so that I do not run unintentionally and overwrite my already scaped json data. These are the results of the web scrape. I saved the results previously so will load the data in again for use.

# player_info_df = get_batters_info(batter_id)

#### Save Final Scape (json)

The below lines are also commented out so that I do not save over my scrape with any of the changes I will make to the dataframe.

# player_info_df = pd.DataFrame(player_info_df)
# player_info_df.to_csv('../data/final_scrape.csv')

#### Load scraped data

player_info_df = pd.read_csv('../data/final_scrape.csv')
player_info_df.drop('Unnamed: 0', axis=1, inplace=True)

player_info_df.head()

final_player_info_df = player_info_df[['batter_id', 'height', 'weight', 'age']]

f'Total number of unique players: {len(final_player_info_df.drop_duplicates())}'

#### Convert all players heights to inches

def height(height_str):
    feet, inches = height_str.split("' ")
    feet = int(feet)
    inches = int(inches[:-1])
    return 12 * feet + inches

final_player_info_df['height'] = final_player_info_df['height'].map(height)

### Save Data

final_player_info_df.to_csv('../data/final_clean_player_info.csv')

##### Now lets explore the data: 02-EDA
